+++ src/main/java/org/realityforge/guiceyloops/server/ServerTestModule.java
@@ -5,6 +5,7 @@ import javax.ejb.SessionContext;
 import javax.enterprise.context.Dependent;
 import javax.enterprise.inject.spi.BeanManager;
 import javax.transaction.TransactionSynchronizationRegistry;
+import java.security.Principal;
 
 /**
  * The client module for all the server side tests in absence of a specific module for a test.
@@ -17,6 +18,7 @@ public final class ServerTestModule
   {
     bindScope( Dependent.class, Scopes.NO_SCOPE );
     bindMock( SessionContext.class );
+    bindMock( Principal.class );
     bind( TransactionSynchronizationRegistry.class ).
       to( TestTransactionSynchronizationRegistry.class ).
       asEagerSingleton();