+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.66 (Pending):
 * Add simplified accessor AbstractServerTest.em(UnitName) for EntityManagers with specific names.
 * Bind mock Principal in ServerTestModule. Useful to get caller principal in CDI applications.
+* Update ServerTestModule to make it possible to override creation of specific resources.
 
 ## 0.65:
 * Remove excessive logging in GreenMailTestModule.